fix: 랭킹 페이지 내 랭킹 반영할 수 있도록 상위 컴포넌트에서 랭킹 정보 조회
